Update requirements.txt
fix: Corrected version specifier for google-generativeai

The requirements file incorrectly used a single equals sign (=)
to specify the version of the google-generativeai package. This
has been corrected to use the double equals sign (==), which is
the correct operator for pinning package versions in pip
requirements files. This ensures that the specified version
(0.8.4) of the google-generativeai library is installed as
intended.